[FIX] hr_holidays_natural_period: Define the correct number of days in some use cases.
Example use case:
- Calendar Monday 16:00-21:00
- Calendar Tuesday-Friday: 09:00-14:00
- Absence from Monday to Sunday in natural day
- Number of days: 7

This is due to the rrule we are using to generate the missing intervals for the non
working days to simulate the natural period. Previously, we use the starting and
ending time, resulting that the daily rrule was omitting the initial or ending day,
as the theoretical hour is after the ending datetime. For the example above, the
rrule will generate the entries from Monday at 16:00, so when the time for the Sunday
rule comes, as Sun 16:00 is bigger than Sun 14:00, which was the ending datetime,
no iteration is done for Sunday.

The solution is to put the rrule with the min time for the starting limit, and the
maximum for the ending one.

It has been checked that there's no problem with timezones due to this patch.

TT55243

[BOT] post-merge updates